Refactor SingleTextViewer tests to one spec file
